requirements.txt
pandas>=1.5.0
numpy>=1.21.0
requests>=2.28.0
python-telegram-bot>=20.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0
schedule>=1.1.0
python-dotenv>=0.19.0
pytest>=7.0.0
pylint>=2.15.0
black>=22.0.0
psycopg2-binary

prometheus_client>=0.16.0
